Identity Server added.

+++ src/AuctionService/Controllers/AuctionController.cs
@@ -3,6 +3,9 @@ using AuctionService.DTOs;
 using AuctionService.Entities;
 using AutoMapper;
 using AutoMapper.QueryableExtensions;
+using Contracts;
+using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -14,12 +17,13 @@ namespace AuctionService.Controllers
     {
         private readonly AuctionDbContext _context;
         private readonly IMapper _mapper;
+        private readonly IPublishEndpoint publishEndpoint;
 
-        public AuctionController(AuctionDbContext auctionDbContext, IMapper mapper)
+        public AuctionController(AuctionDbContext auctionDbContext, IMapper mapper,IPublishEndpoint publishEndpoint)
         {
             _context = auctionDbContext;
             _mapper = mapper;
-
+            this.publishEndpoint = publishEndpoint;
         }
 
         [HttpGet]
@@ -47,37 +51,44 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDTO>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDTO>> CreateAuction(CreateAuctionDTO createAuctionDTO)
         {
             var auction = _mapper.Map<Auction>(createAuctionDTO);
-            // TODO: add current user as seller.
-            auction.Seller = "test";
+           
+            auction.Seller = User.Identity.Name;
             _context.Auctions.Add(auction);
 
+            var toBePublishedAuction = _mapper.Map<AuctionDTO>(auction);
+            await publishEndpoint.Publish(_mapper.Map<AuctionCreated>(toBePublishedAuction));
+
             var result = await _context.SaveChangesAsync() > 0;
+
             if (!result)
             {
                 return BadRequest("Could not save changes to DB");
             }
 
-            return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, _mapper.Map<AuctionDTO>(auction));
+            return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, toBePublishedAuction);
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDTO updateAuctionDTO)
         {
             var auction = await _context.Auctions.Include(x => x.Item).FirstOrDefaultAsync(x => x.Id == id);
 
             if (auction == null) return NotFound();
-
-            // TODO: check seller==username
+            if (auction.Seller != User.Identity.Name) return Forbid();
 
             auction.Item.Make = updateAuctionDTO.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDTO.Model ?? auction.Item.Model;
             auction.Item.Color = updateAuctionDTO.Color ?? auction.Item.Color;
             auction.Item.Mileage = updateAuctionDTO.Mileage ?? auction.Item.Mileage;
             auction.Item.Year = updateAuctionDTO.Year ?? auction.Item.Year;
+
+            await publishEndpoint.Publish(_mapper.Map<AuctionUpdated>(auction));
 
             var result = await _context.SaveChangesAsync() > 0;
 
@@ -86,6 +97,7 @@ namespace AuctionService.Controllers
             return BadRequest("Problem occured while updating auction!");
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
@@ -93,9 +105,12 @@ namespace AuctionService.Controllers
 
             if (auction == null) return NotFound();
 
-            // TODO: check seller = username 
+            if (auction.Seller != User.Identity.Name) return Forbid();
 
             _context.Auctions.Remove(auction);
+
+            await publishEndpoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString()});
+
             var result = await _context.SaveChangesAsync() > 0;
 
             if (!result) return BadRequest("Auction could not be deleted!");
